refactor(landscape dialog) Move LandscapeModal to dialog, show it with flag
