--- src/lib/ui/core/LandscapeModal/index.ts
@@ -1,1 +1,0 @@
-export { showLandscapeModal$ } from './LandscapeModal.svelte'
