--- src/stories/Design System - Core UI/LandscapeModal/index.stories.ts
@@ -1,14 +1,0 @@
-import type { Meta, StoryObj } from '@storybook/svelte'
-import component from './index.svelte'
-
-const meta = {
-  component,
-  parameters: {
-    layout: 'fullscreen',
-  },
-} satisfies Meta<component>
-type Story = StoryObj<typeof meta>
-
-export default meta
-
-export const Dialog: Story = {}
